New particles which you get points for collecting.
So far just one per brick, and they never die as long as you stay under
them. They also render weirdly if they go off the sides.

Needed to edit Draw.hx because the comparison fails: it would work if we
used UInts, but it's not obvious that the comparison makes things faster,
let alone enough faster to be worth bothering to fix.