Add death sound.
We no longer reset the ball when we lose, which is probably better, but
it does get stuck at the bottom, which is bad. Ideally we probably want
the paddle to drop.

What was previously happening was: when we lose our last ball, we play
"lose" and start fading out. Next frame, that "if" condition is still
true so we call placeBall().

If we still did that, then the death sound would play at the same time
as "lose", which is bad. So we don't do it any more.